Use two-tag method for sniper angle ticks
The sniper HUD ticks are offset from the top left based on the default safe zone settings. Safe zones can be changed with commands, and if you do this it will break the exact same way as Halo PC, so here we use the same fix so the safe zones can be changed.